imap: move over imapclient.Search{Options,Criteria}
